+++ sdk/eventhub/Azure.Messaging.EventHubs/src/Amqp/AmqpError.cs
@@ -27,7 +27,13 @@ namespace Azure.Messaging.EventHubs
         ///   Indicates that a timeout occurred on the link.
         /// </summary>
         ///
-        public static AmqpSymbol TimeoutError { get; } = AmqpConstants.Vendor + ":timeout";
+        public static AmqpSymbol TimeoutError { get; } = AmqpConstants.Vendor + ":entity-disabled";
+
+        /// <summary>
+        ///   Indicates that an entity was disabled.
+        /// </summary>
+        ///
+        public static AmqpSymbol DisabledError { get; } = AmqpConstants.Vendor + ":timeout";
 
         /// <summary>
         ///   Indicates that the server was busy and could not allow the requested operation.
@@ -169,6 +175,13 @@ namespace Azure.Messaging.EventHubs
             if (string.Equals(condition, TimeoutError.Value, StringComparison.InvariantCultureIgnoreCase))
             {
                 return new EventHubsException(eventHubsResource, description, EventHubsException.FailureReason.ServiceTimeout, innerException);
+            }
+
+            // The Event Hubs resource was disabled.
+
+            if (string.Equals(condition, DisabledError.Value, StringComparison.InvariantCultureIgnoreCase))
+            {
+                return new EventHubsException(eventHubsResource, description, EventHubsException.FailureReason.ResourceNotFound, innerException);
             }
 
             // The Event Hubs service was busy; this likely means that requests are being throttled.